Support cropping in the video output
This changes the IVideoPlayer interface to pass two additional
parameters - the width/height parameters are kept as the full
buffer size, with two additional parameters specifying how much of
it is supposed to be visible. This removes the duplicated alignment
calculation, making sure it's done in one place only - in the calling
vout code.